requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
filelock==3.12.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
Pillow==9.5.0
requests==2.30.0
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.5.0
urllib3==2.0.2
tqdm==4.65.0
aiohttp==3.8.4
loguru==0.7.0
munch==3.0.0
pycryptodome==3.18.0
simplejson==3.19.1
yarl==1.9.2